+++ includes/class-data-science-page.php
@@ -237,7 +237,7 @@ class EDIT_Data_Science_Page {
                             <div class="container">
                                 <div class="row">
                                     <?php foreach ( $slugs as $slug ) {
-                                        echo EDIT_Pillar_Courses::render_card( $slug );
+                                        echo EDIT_Pillar_Courses::render_card( $slug, $group );
                                     } ?>
                                 </div>
                             </div>