+++ includes/class-marketing-digital-page.php
@@ -316,7 +316,7 @@ class EDIT_Marketing_Digital_Page {
                             <div class="container">
                                 <div class="row">
                                     <?php foreach ( $slugs as $slug ) {
-                                        echo EDIT_Pillar_Courses::render_card( $slug );
+                                        echo EDIT_Pillar_Courses::render_card( $slug, $group );
                                     } ?>
                                 </div>
                             </div>