+++ includes/class-ux-ui-design-page.php
@@ -248,7 +248,7 @@ class EDIT_UX_UI_Design_Page {
                             <div class="container">
                                 <div class="row">
                                     <?php foreach ( $slugs as $slug ) {
-                                        echo EDIT_Pillar_Courses::render_card( $slug );
+                                        echo EDIT_Pillar_Courses::render_card( $slug, $group );
                                     } ?>
                                 </div>
                             </div>